requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-decouple
django-rest-swagger==2.1.2
drf-spectacular
psycopg2-binary>=2.9
django-filter

django-debug-toolbar
faker